+++ include/Shard.hpp
@@ -7,7 +7,6 @@
 #include <string>
 #include <unordered_map>
 #include <vector>
-#include "worker.h"
 
 class Shard {
    private:
